Allow plugin methods to override default methods and call super
Currently, because all class and instance methods are defined
directly on Cuba, you cannot override them in plugins.  There
are two ways to fix this.  One is to require ruby 2.0 and use
Module.prepend.  The second is moving the methods to a plugin
itself, leave the Cuba class empty, and just load that plugin
into Cuba.  This takes the latter approach, which is also the
approach that Sequel uses.