Attempt to look up non-existing settings in parent classes (see #21)
This fixes an issue where settings are added to Cuba.settings *after* inheriting
from `Cuba` are not picked up by the child classes. The problem is described in
issue #21.

If a setting does not exist, Settings are looked up in the settings Hash of the parent class using `Hash#default_proc`. The settings Hash is still deepcloned, so
settings are still overridable.